+++ src/GlyphFace.cpp
@@ -33,16 +33,16 @@ int32 GlyphFace::getMetric(uint8 metric) const
 {
     switch (metrics(metric))
     {
-        case kgmetLsb       : return m_bbox.bl.x;
-        case kgmetRsb       : return m_advance.x - m_bbox.tr.x;
-        case kgmetBbTop     : return m_bbox.tr.y;
-        case kgmetBbBottom  : return m_bbox.bl.y;
-        case kgmetBbLeft    : return m_bbox.bl.x;
-        case kgmetBbRight   : return m_bbox.tr.x;
-        case kgmetBbHeight  : return m_bbox.tr.y - m_bbox.bl.y;
-        case kgmetBbWidth   : return m_bbox.tr.x - m_bbox.bl.x;
-        case kgmetAdvWidth  : return m_advance.x;
-        case kgmetAdvHeight : return m_advance.y;
+        case kgmetLsb       : return int32(m_bbox.bl.x);
+        case kgmetRsb       : return int32(m_advance.x - m_bbox.tr.x);
+        case kgmetBbTop     : return int32(m_bbox.tr.y);
+        case kgmetBbBottom  : return int32(m_bbox.bl.y);
+        case kgmetBbLeft    : return int32(m_bbox.bl.x);
+        case kgmetBbRight   : return int32(m_bbox.tr.x);
+        case kgmetBbHeight  : return int32(m_bbox.tr.y - m_bbox.bl.y);
+        case kgmetBbWidth   : return int32(m_bbox.tr.x - m_bbox.bl.x);
+        case kgmetAdvWidth  : return int32(m_advance.x);
+        case kgmetAdvHeight : return int32(m_advance.y);
         default : return 0;
     }
 }
